requirements.txt
aiogram>=3.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0